Add the "Visibility" column to the cloud scores list
