+++ src/cloud/CMakeLists.txt
@@ -30,10 +30,13 @@ set(MODULE_SRC
     ${CMAKE_CURRENT_LIST_DIR}/iauthorizationservice.h
     ${CMAKE_CURRENT_LIST_DIR}/icloudconfiguration.h
     ${CMAKE_CURRENT_LIST_DIR}/cloudtypes.h
+    ${CMAKE_CURRENT_LIST_DIR}/cloudqmltypes.h
     ${CMAKE_CURRENT_LIST_DIR}/clouderrors.h
+
+    ${CMAKE_CURRENT_LIST_DIR}/internal/cloudconfiguration.h
+    ${CMAKE_CURRENT_LIST_DIR}/internal/cloudconfiguration.cpp
     ${CMAKE_CURRENT_LIST_DIR}/internal/oauthhttpserverreplyhandler.h
     ${CMAKE_CURRENT_LIST_DIR}/internal/oauthhttpserverreplyhandler.cpp
-
     ${CMAKE_CURRENT_LIST_DIR}/internal/abstractcloudservice.h
     ${CMAKE_CURRENT_LIST_DIR}/internal/abstractcloudservice.cpp
     ${CMAKE_CURRENT_LIST_DIR}/musescorecom/imusescorecomservice.h
@@ -42,9 +45,6 @@ set(MODULE_SRC
     ${CMAKE_CURRENT_LIST_DIR}/audiocom/iaudiocomservice.h
     ${CMAKE_CURRENT_LIST_DIR}/audiocom/audiocomservice.h
     ${CMAKE_CURRENT_LIST_DIR}/audiocom/audiocomservice.cpp
-
-    ${CMAKE_CURRENT_LIST_DIR}/internal/cloudconfiguration.h
-    ${CMAKE_CURRENT_LIST_DIR}/internal/cloudconfiguration.cpp
 
     ${CMAKE_CURRENT_LIST_DIR}/view/cloudsmodel.h
     ${CMAKE_CURRENT_LIST_DIR}/view/cloudsmodel.cpp
