+++ src/cloud/cloudmodule.cpp
@@ -33,6 +33,8 @@
 #include "view/cloudsmodel.h"
 #include "view/musescorecomauthorizationmodel.h"
 
+#include "cloudqmltypes.h"
+
 using namespace mu::cloud;
 using namespace mu::modularity;
 
@@ -74,6 +76,9 @@ void CloudModule::registerUiTypes()
 {
     qmlRegisterType<CloudsModel>("MuseScore.Cloud", 1, 0, "CloudsModel");
     qmlRegisterType<MuseScoreComAuthorizationModel>("MuseScore.Cloud", 1, 0, "MuseScoreComAuthorizationModel");
+
+    qmlRegisterUncreatableType<QMLCloudVisibility>("MuseScore.Cloud", 1, 0, "CloudVisibility",
+                                                   "Not creatable as it is an enum type");
 
     modularity::ioc()->resolve<ui::IUiEngine>(moduleName())->addSourceImportPath(cloud_QML_IMPORT);
 }
